moved orders GetOne, UpdateOne, DeleteOne, route logic to controllers
